TG BOT/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
SQLAlchemy>=2.0.0
aiosqlite>=0.17.0
aiohttp>=3.8.0
python-dateutil>=2.8.2 